Split out object interface. Add support for the speaker API. Add sample script to test speaker API on staging.
